Rename condition -> conditions
