+++ bloop/conditions.py
