+++ bloop/tracking.py
@@ -1,4 +1,4 @@
-from .condition import Condition
+from .conditions import Condition
 from .util import WeakDefaultDictionary, signal
 
 # Tracks the state of instances of models:
