+++ tests/unit/test_column.py
@@ -1,7 +1,7 @@
 import operator
 
 import pytest
-from bloop.condition import AttributeExists, BeginsWith, Between, Contains, In
+from bloop.conditions import AttributeExists, BeginsWith, Between, Contains, In
 from bloop.models import Column
 from bloop.types import Integer
