+++ tests/unit/test_condition.py
@@ -2,7 +2,7 @@ import operator
 import uuid
 
 import pytest
-from bloop.condition import (
+from bloop.conditions import (
     And,
     AttributeExists,
     BeginsWith,
